ingestion/requirements.txt
supabase==2.10.0
pandas==2.2.3
requests==2.32.3
openpyxl==3.1.5
open-mastr==0.17.1
python-dotenv==1.0.1
pyproj==3.6.1
python-dateutil==2.9.0
feedparser==6.0.11
yfinance==0.2.50
lxml==5.3.0
